refactor: rename useAudio hook module
